ajuste e criacao do elemento de HomeEvents para o page da home
